+++ .gitignore
@@ -4,6 +4,9 @@ target
 *.tmp
 .idea
 
+**/*.output/*
+!**/*.output/.gitkeep
+
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
