+++ sources/libs/bisque-cipher/samples/decrypted/image.png
