+++ sources/libs/bisque-cipher/samples/encrypted/empty.cbc
