+++ sources/libs/bisque-cipher/samples/encrypted/image.cbc
