+++ sources/libs/bisque-cipher/samples/encrypted/text_larger_than_block_size.cbc
