+++ sources/libs/bisque-cipher/samples/encrypted/text_smaller_than_block_size.cbc
